requirements.txt
minimalmodbus==2.1.1
pyModbusTCP==0.3.0
pyserial==3.5